refactor(server): split team.ts — extract handleBuildTeam (S27.8.27)
Sixth slice on `apps/server/src/routes/team.ts`. Extract the team
builder into `routes/team-build-handler.ts` (320 lines, under DoD
400).

Endpoint covered:
- `POST /build` — handleBuildTeam (atomic team creation from
  builder: positions + count + budget + Star Players + staff)

Cascade validation (roster, min/max per position, 11-16 players,
Star Player pairs, 16 cap, total budget). Creates team +
teamPlayer[] + teamStarPlayer[] + recomputes TV.

Re-export from `team.ts` to preserve the public API consumed by
`team.test.ts`.

Leaf dependencies only (`prisma`, `sendError/sendSuccess`,
`updateTeamValues`, types/helpers from `@bb/game-engine`,
`validateStarPlayerPairs`/`calculateStarPlayersCost`/
`validateStarPlayersForTeam`, `getRosterFromDb`, `resolveRuleset`,
`serverLog`). No cycle.

`team.ts` shrinks 1065 -> 852 lines (-213 net lines, cumulative
-1562 vs origin 2414 since S27.8.22 = -65%).

3 new smoke tests `team-build-handler.test.ts`:
- exports x2 (named export + re-export from team.ts)
- 400 gate on unknown roster

- 1827/1827 server tests pass (was 1824, +3 new)
- typecheck clean
